Fixing old backport for content-disposition in reporting
